Refactored the TreeNode::Node class, removed define_methods
